% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{get_brand}
\alias{get_brand}
\title{Load Quarto branding elements into global environment}
\usage{
get_brand(root_dir = ".", brand_filename = "brand.yml")
}
\arguments{
\item{root_dir}{Root directory to start search.}

\item{brand_filename}{Name of the branding YAML file.}
}
\value{
Invisible NULL. Sets brand/brand_colors in global environment.
}
\description{
Searches for a brand.yml file and loads branding information.
}
\examples{
get_brand()
}
